refactor: rename pyneuromatic/analysis/ to pyneuromatic/tools/
Renames the analysis module to tools throughout the codebase:

- pyneuromatic/analysis/ → pyneuromatic/tools/
- tests/test_analysis/ → tests/test_tools/
- All imports updated from pyneuromatic.analysis to pyneuromatic.tools

No logic changes.

Closes #306
