--- pyneuromatic/analysis/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-from pyneuromatic.analysis.nm_tool_event import NMToolEvent, NMToolEventConfig
-from pyneuromatic.analysis.nm_tool_fit import NMToolFit, NMToolFitConfig
-from pyneuromatic.analysis.nm_tool_main import NMToolMain
-from pyneuromatic.analysis.nm_pulse import NMPulse, NMPulseContainer
-from pyneuromatic.analysis.nm_tool_pulse import NMToolPulse, NMToolPulseConfig
-from pyneuromatic.analysis.nm_tool_spike import NMToolSpike, NMToolSpikeConfig
-from pyneuromatic.analysis.nm_tool_stats import NMToolStats, NMToolStatsConfig
