+++ pyneuromatic/tools/nm_tool_main.py
@@ -20,8 +20,8 @@ Paper: https://doi.org/10.3389/fninf.2018.00014
 """
 from __future__ import annotations
 
-from pyneuromatic.analysis.nm_main_op import NMMainOp, NMMainOpAverage, op_from_name
-from pyneuromatic.analysis.nm_tool import NMTool
+from pyneuromatic.tools.nm_main_op import NMMainOp, NMMainOpAverage, op_from_name
+from pyneuromatic.tools.nm_tool import NMTool
 from pyneuromatic.core.nm_folder import NMFolder
 import pyneuromatic.core.nm_history as nmh
 import pyneuromatic.core.nm_utilities as nmu
