+++ pyneuromatic/tools/nm_tool_stats.py
@@ -24,11 +24,11 @@ from typing import Any
 
 import numpy as np
 
-from pyneuromatic.analysis.nm_stat_utilities import stat
-from pyneuromatic.analysis.nm_stat_win import NMStatWinContainer  # noqa: F401
-from pyneuromatic.analysis.nm_tool import NMTool
-from pyneuromatic.analysis.nm_tool_config import NMToolConfig
-from pyneuromatic.analysis.nm_tool_folder import NMToolFolder
+from pyneuromatic.tools.nm_stat_utilities import stat
+from pyneuromatic.tools.nm_stat_win import NMStatWinContainer  # noqa: F401
+from pyneuromatic.tools.nm_tool import NMTool
+from pyneuromatic.tools.nm_tool_config import NMToolConfig
+from pyneuromatic.tools.nm_tool_folder import NMToolFolder
 from pyneuromatic.core.nm_data import NMData
 from pyneuromatic.core.nm_dataseries import NMDataSeries
 from pyneuromatic.core.nm_epoch import NMEpoch
